Rearranged files, added license headers
